requirements.txt
selenium
pandas
webdriver-manager
python-dotenv
